reduce trie depth and handle more edge cases
In the previous version, when a trie was constructed, we made sure all paths were prefixed with "/". Since we then immediately split on "/", this would mean that the depth of the trie was always 1 more than it needed to be, since the root was always empty (per the design of tries), and its child was always also empty, since it was all of the characrters before "/".