reduce initial trie depth
The root of the trie is always empty, but due to passing in leading slashes, we always ended up with a leading slash that was split. This resulted in the first element always being a single node with an empty value and all of the other children, effectively making it a copied root.