switch to only including TRACE if the server supports it
The amount of requests that are OPTIONS or method not allowed is likely quite small versus the total number of requests a server is expected to serve.
Also, most servers will choose to disable TRACE requests if they go to production.
The combination of these means its much better to just assume TRACE is _not_ enabled for the server, and then have the middleware add it in if needed, since this action will occur much much much less frequently than needing to check its presence.
In doing this, we dont need to register this middleware unless the TRACE method is enabled, meaning it will likely very rarely be used.