use `RawPath` instead of `Path` in default 404
This makes it more explicit and also makes sure that rewritten URIs use their edge path, instead of the rewritten path. This is important as it reflects what the client actually sent to the server. The edge, rewritten and raw paths are still logged together for the 404 response by default so debugging shows us what path the actual request was routed with