move routing to before all middleware
There are some situations where we might want to have a routed request in the middleware. For example, if we want to extract path parameters from the URI. Previously, this would have required doing additional work by using regex or similar to extract known path parameters inside middleware to be able to use them. Instead, this moves routing to before all middleware, meaning the middleware can treat the request as routed and act accordingly. Currently, even if a request cannot be routed, it is still processed through all middleware before ultimately exiting. This might change depending on how I feel it goes, otherwise we could introduce a config option