make `realChain.last` be a `HandlerFunc`
The last elemenet of the middleware should not be able to pass forward to the next piece of middleware - this could cause issues due to there being no next. Since we control what the last piece of middleware is (its always the core handler that handles the request and performs specific logic, such as enforcing client `Accept` headers), we know that the chain is not proceeded within it. However, for increased security within the code, this change enforces that behaviour, meaning we are confirmed to have a terminal middleware.