make error comparison more straightforward
Within `errors.Is`, the errors package will check for the presence of an `Is(error) bool` method on the error type, which can be used for more complex error types (i.e. non-value types). This commit implements this method to use a status code comparison for routeit error types. This is mainly needed for tests, but will also help elsewhere, for example in dedicated error handling