remove boolean from `Request.PathParam`
In practice I did not find it useful to have this boolean, and due to how dynamic routing works, we can expect it to return a non-empty value all the time, unless the name does not match any named segments. The empty string does enough here to infer if it is present or not due to the strict requirement on > 0 characters in a path segment.