fix issue with updates
MySQL optimises `UPDATE` queries to return "no rows affected" if the query does not change any rows. So if we receive a request to update an item (whose name is "Bread") to update its name to "Bread", MySQL will tell us no rows have been affected, even though the query executed successfully. By contrast, if the request told us to update the name to "Cheese", this would return "1 row affected". This was causing the update endpoints to sometimes return `404: Not Found` incorrectly. The solution is to remove the "rows affected" check altogether, which also simplifies some tests and the repositories in general.