Store request IDs (#35)
### Summary
<!-- A high level summary of the changes, including any gotchas that the
reviewer should watch out for when reviewing. -->

This PR allows inclusion of ID's on each incoming request. This is done
through a piece of middleware that is conditionally installed to the
server. We also include the request ID in a response header, which is
defaulted to `"X-Request-Id"` but can be changed.

### Motivation
<!-- Why is this change necessary? This can be a link to a GitHub issue
or reproduction steps for a bug etc. -->

This improves debugging by adding allowing users to add identifiers to
incoming requests.

### Test plan
<!-- How did you test the changes? Are there edge cases you are missing
or were unable to test? -->

- [x] Unit tests

This is not used anywhere yet but will be once the next version bump is
published.