[https] Refactor socket to internal interface (#39)
### Summary
<!-- A high level summary of the changes, including any gotchas that the
reviewer should watch out for when reviewing. -->

This PR introduces a new internal package `internal/socket` and a
`Socket` interface. The interface is used to establish server-side
connections using network sockets. A `Bind`, `Serve` and `Close` method
are all exposed on the interface.

An initial concrete implementation of a TCP socket is added and
connected to the server's socket implementation. This uses callbacks to
manage how connections are consumed and is thread-safe.

### Motivation
<!-- Why is this change necessary? This can be a link to a GitHub issue
or reproduction steps for a bug etc. -->

I am exploring adding HTTPS support to the project. In order to do this
we will need to support TCP, TLS over TCP and a combination of both. To
keep the server's implementation easier to work with, we can abstract
them behind the internal `socket` package.

### Test plan
<!-- How did you test the changes? Are there edge cases you are missing
or were unable to test? -->

- [x] Unit tests
- [x] Manual tests locally