[https] TLS over TCP socket (#40)
### Summary
<!-- A high level summary of the changes, including any gotchas that the
reviewer should watch out for when reviewing. -->

Following on from the previous PR, this PR adds a concrete TLS over TCP
implementation of the `Socket` interface. This uses the `crypto/tls`
socket and is similar in most implementation details to the existing TCP
socket.

### Motivation
<!-- Why is this change necessary? This can be a link to a GitHub issue
or reproduction steps for a bug etc. -->

In order to support HTTPS, we need to be able to support responding to
TLS connections, which also includes the TLS handshake. Once decrypted,
the message should be a HTTP/1.1 message, so can be reacted to in the
same way the server normally would.

### Test plan
<!-- How did you test the changes? Are there edge cases you are missing
or were unable to test? -->

- [x] Unit tests