[https] Combined socket for TLS and TCP (#41)
### Summary
<!-- A high level summary of the changes, including any gotchas that the
reviewer should watch out for when reviewing. -->

This PR combines the two `Socket` implementations for plain TCP and TLS
over TCP into a combined socket. This allows a server to listen (on
different ports) to both HTTP and HTTPS communications, which many
servers commonly do. Individual TCP and TLS over TCP sockets are bound
and spun up when the combined socket begins listening. Since
`Socket.Serve` is typically blocking, the two invocations are put in
separate go-routines. To maintain consistency, we use channels to block
the new implementation of `Socket.Serve`, which will continue to block
until both sockets close (and therefore terminate their own
implementations of `Socket.Serve`). We can rely on the two underlying
sockets being thread-safe, so don't have any explicit thread-safe
behaviour or mechanisms in this implementation.

### Motivation
<!-- Why is this change necessary? This can be a link to a GitHub issue
or reproduction steps for a bug etc. -->

A user may want to accept both HTTP and HTTPS messages. In that case, we
need two separate ports and listeners. By implementing the `Socket`
interface, we would not need to change any of the server's code (apart
from initialisation) to support listening on both ports.

### Test plan
<!-- How did you test the changes? Are there edge cases you are missing
or were unable to test? -->

- [x] Unit tests