[https] Server config for TLS (#42)
### Summary
<!-- A high level summary of the changes, including any gotchas that the
reviewer should watch out for when reviewing. -->

This PR refactors server configuration to allow the user to specify a
number of combinations for which transport the server uses. There are
three main buckets
- HTTP only
- HTTPS only
- Both HTTP and HTTPS

The `ServerConfig.Port` property is removed in favour of a more
descriptive `HttpConfig` struct, which contains a HTTP port, HTTPS port
and TLS config. Whenever a TLS config is provided, we know we are
dealing with at least HTTPS. In such cases, we default to serving HTTPS
on port 443 unless specified and do not serve HTTP content unless the
user explicitly configures a HTTP port. If `HttpConfig` is left empty,
we default to serving HTTP only content over port 8080 (the previous
default). If just a HTTP port is provided, we listen for plain HTTP
content on that port.

Necessary guard-rails (such as attempting to use HTTPS without a TLS
config) are enforced.

At server initialisation, we figure out the corresponding ports and
determine which `internal/socket.Socket` to use. This is then stored on
the server and used whenever it is started up.

> [!NOTE]
> Technically the removal of `ServerConfig.Port` is
backwards-incompatible. However, I am the only user of the library, and
the changes are not used anywhere in production. As such this is a safe
change

### Motivation
<!-- Why is this change necessary? This can be a link to a GitHub issue
or reproduction steps for a bug etc. -->

This allows us to build on the socket implementations introduced
previously. We can now support different combinations of HTTP(S) on the
server easily.

### Test plan
<!-- How did you test the changes? Are there edge cases you are missing
or were unable to test? -->

- [x] Existing tests, some refactored
- [x] Local testing