fail faster in tests when value is not found
Some of the library tests dont fail early enough, leading to panics in the test code (for example whenever the length of the headers is 0 and we then attempt to access index 0). This causes confusion in the tests and makes it harder to understand the point of failure. By failing hard when we know that the test wont succeed and could panic, we surface the error message more easily to the user