[https] Example server accepting both HTTP and HTTPS (#43)
### Summary
<!-- A high level summary of the changes, including any gotchas that the
reviewer should watch out for when reviewing. -->

This PR adds the first example project that uses HTTPS servers. This
includes a sample certificate and private key which the example uses,
and future examples will use for ease. This reduces complication in
setup and the key and certs are perfectly safe as they are not used
outside of the example projects in the repo.

The new example project contains 1 endpoint and accepts communication
over both HTTP and HTTPS. When the request is made using TLS (and
therefore HTTPS), we thank the user for their increased security.

### Motivation
<!-- Why is this change necessary? This can be a link to a GitHub issue
or reproduction steps for a bug etc. -->

To validate how HTTPS connections with `routeit` work and show an
example of setting one up.

### Test plan
<!-- How did you test the changes? Are there edge cases you are missing
or were unable to test? -->

- [x] E2E test
- [x] Manual tests locally