[https] Allow test client to be TLS aware (#44)
### Summary
<!-- A high level summary of the changes, including any gotchas that the
reviewer should watch out for when reviewing. -->

This PR updates the creation of the `TestClient` to allow users to
create a TLS-aware test client. The TLS connection state provided is
passed into the server to handle a request in E2E tests. We do the same
in unit tests for handlers and middleware, where the
`TestRequestOptions` struct is extended to also include TLS connection
state.

### Motivation
<!-- Why is this change necessary? This can be a link to a GitHub issue
or reproduction steps for a bug etc. -->

Some handlers and middleware are TLS-aware - i.e. they may reject a
request if it is not using TLS or decide to only expose certain
information etc. Previously there was no way to test this automatically,
as there was no mechanism to inject TLS state into either unit or E2E
tests.

### Test plan
<!-- How did you test the changes? Are there edge cases you are missing
or were unable to test? -->

- [x] Existing