[https] HTTPS only example server (#45)
### Summary
<!-- A high level summary of the changes, including any gotchas that the
reviewer should watch out for when reviewing. -->

This PR introduces a new example server that only accepts HTTPS
requests. Plain HTTP requests will not reach the server logic and will
be closed due to the server expecting a TLS handshake after the TCP
handshake.

The server exposes 1 endpoint that shows some details about the client
and server's TLS handshake.

### Motivation
<!-- Why is this change necessary? This can be a link to a GitHub issue
or reproduction steps for a bug etc. -->

To show how to make a server only accept HTTPS communication, as well as
verifying that the TLS connection state stored on the request is
correct.

### Test plan
<!-- How did you test the changes? Are there edge cases you are missing
or were unable to test? -->

- [x] New E2E test
- [x] Local testing using cURL