[https] Middleware to upgrade HTTP to HTTPS (#46)
### Summary
<!-- A high level summary of the changes, including any gotchas that the
reviewer should watch out for when reviewing. -->

This PR adds config options and middleware to upgrade HTTP connections
to HTTPS.

When the server receives a HTTP connection, it can redirect to the
corresponding HTTPS URI. Most clients will follow this redirect, or at
least surface it to the user. From there, they will establish the TLS
handshake and make a HTTPS request. When the server receives a HTTPS
request, they can also inform the client to keep using that scheme for
future requests. Using a `max-age` value in a header, the server can
inform the client to cache the instruction to use HTTPS for a specific
amount of time. Some clients don't respect this, but all browsers do.

The enablement of this config option is done by introducing two new
fields - `UpgradeToHttps` and `UpgradeInstructionMaxAge` to
`ServerConfig.HttpConfig`. These are validated properly (such as
ensuring they are not enabled when the server is not enabled for HTTPS),
and also mean that we know the server will be responding to both HTTP
and HTTPS communication. So, unlike in previous scenarios, if the user
doesn't provide a HTTP port to listen on, we listen on the default HTTP
port of 80 if these config options are enabled.

### Motivation
<!-- Why is this change necessary? This can be a link to a GitHub issue
or reproduction steps for a bug etc. -->

Some servers wish to receive both HTTP and HTTPS traffic, but only want
to process HTTPS traffic. They want traffic over HTTP to be instructed
to retry using HTTPS for security purposes. `routeit` already allows
creating a server that listens for both HTTP and HTTPS traffic, so it
seems sensible that this behaviour could also be controllable.

### Test plan
<!-- How did you test the changes? Are there edge cases you are missing
or were unable to test? -->

- [x] Unit tests (new + old)
- [x] Local tests