[https] Example for upgrading HTTP to HTTPS (#47)
### Summary
<!-- A high level summary of the changes, including any gotchas that the
reviewer should watch out for when reviewing. -->

This PR adds a simple example app that showcases how `routeit`
automatically handles upgrading connections from HTTP to HTTPS if
enabled on the server.

### Motivation
<!-- Why is this change necessary? This can be a link to a GitHub issue
or reproduction steps for a bug etc. -->

Validation of the concept and a demonstration of it.

### Test plan
<!-- How did you test the changes? Are there edge cases you are missing
or were unable to test? -->

- [x] E2E tests
- [x] Local testing