[chore] Bump examples to v1.2.0 (#48)
### Summary
<!-- A high level summary of the changes, including any gotchas that the
reviewer should watch out for when reviewing. -->

This PR bumps the example projects from v1.1.0 to v1.2.0. Additionally,
the `http` examples that took a dependency on a local replacement of the
`routeit` package have been updated to now use a published version of
the package. The script to bump the examples has also been updated to
allow for more consistent behaviour around empty arrays.

Lastly, the `middleware/cors`, `simple`, `static/rewrites/static` and
`todo` examples have all been patched to use the new `HttpConfig` struct
instead of `Port`.

### Motivation
<!-- Why is this change necessary? This can be a link to a GitHub issue
or reproduction steps for a bug etc. -->

Keep examples up to date.

### Test plan
<!-- How did you test the changes? Are there edge cases you are missing
or were unable to test? -->

- [x] Existing E2E tests